// Ionic4/src/app/shared/models/inspection-building-course.ts
import {InspectionBuildingCourseLane} from './inspection-building-course-lane';

export class InspectionBuildingCourse {
  id: string;
  idBuilding: string;
  idFirestation: string;

  isActive: boolean = true;
  hasBeenModified: boolean = false;

  lanes: InspectionBuildingCourseLane[];
}